Refactor IsoObject to IsoEntity and enhance components
Renamed IsoObject to IsoEntity and IsoObjectManager to IsoEntityManager, updating all references accordingly. Added IsoTransformComponent and FTransform for entity transforms. Improved component and entity management, including new methods for component retrieval and entity creation. Enhanced logging, tile direction handling, and tile rendering to support flipping. Updated camera to follow entity transforms and improved update logic throughout.

+++ Engine/Core/Components/Objects/ObjectComponents/ComponentTick.cs
@@ -4,16 +4,16 @@ namespace Iso.Engine.Core.Components.IsoObjectComponents
 {
     public class IsoComponentTick
     {
-        public IsoComponentTick(Action<float> thisTickFunction) 
+        public IsoComponentTick(Action<double> thisTickFunction) 
         {
             tickFunction = thisTickFunction;
         }
 
         public bool canEverTick = true;
         public bool runtimeTickEnabled = true;
-        public Action<float> tickFunction = null;
+        public Action<double> tickFunction = null;
 
-        public void TickComponent(float delta)
+        public void TickComponent(double delta)
         {
             if(runtimeTickEnabled)
             {
